+++ lib/formular/helper.rb
@@ -22,17 +22,20 @@ module Formular
 
     class << self
       def _builder
-        @builder || load_builder(:basic)
+        @builder || :basic
       end
       attr_writer :builder
 
       def builder(name)
-        self.builder = load_builder(name)
+        self.builder = name
       end
 
       def load_builder(name)
+        builder_const = BUILDERS.fetch(name, nil)
+        return name unless builder_const
+
         require "formular/builders/#{name}"
-        Formular::Builders.const_get(BUILDERS.fetch(name))
+        Formular::Builders.const_get(builder_const)
       end
     end
 
@@ -40,8 +43,10 @@ module Formular
 
     def builder(model, **options)
       builder_name = options.delete(:builder)
+      builder_name ||= Formular::Helper._builder
 
-      builder = builder_name ? Formular::Helper.load_builder(builder_name) : Formular::Helper._builder
+      builder = Formular::Helper.load_builder(builder_name)
+
       options[:model] ||= model
 
       builder.new(options)