+++ CHANGELOG.md
@@ -1,12 +1,34 @@
 # Master
+### Added
 
+* Updated bootstrap 3 and 4 buttons to accept the following options :size, :block, :color. These will ensure the correct classes are applied to the element (without causing conflicts with defaults).
+
+### Internal
+
+* [BREAKING] Swap attribute handling to be a whitelist rather than a blacklist
+
+
+# v0.2.3 2017-12-08
 ### Fixed
+
+* Improve Escape html - use for option values and optgroup labels
+* Multi Select elements now mark options as selected
+
+# v0.2.2 2017-09-13
+
+### Added
 
 * Select elements prompt and include_blank options now operate the same way as simple form
 * added procces_option to the Element API
-* Escape html (controls, labels errors hints)
-* provide an element module for easily escaping html values
+* Checkboxes now support `checked_value` option as an alias of `value`. Same as SimpleForm.
+* you can now set your own builders as the default in the helper
+
+### Fixed
+
+* Escape html (controls, labels errors hints) (fixes #29)
 * setting a default builder in the helper now actually works!
+* Select elements appends array signifier (`[]`) to element name attribute when `multiple: true` (fixes #40)
+* options should always override default values, even when nil. (fixes #39)
 
 ### Internal
 
@@ -16,6 +38,7 @@
 * added a module for html_escape
 * renamed WrappedControl module to Wrapped and stopped including control
 * changes the order of default_hash to better respect inheritance ordering
+* provide an element module for easily escaping html values
 
 # v0.2.1 2016-09-29